reorganize utility methods
